// SDKs/twilio.com/twilio_chat_v2/1.36.0/typescript/src/sdk/models/shared/chatv2servicechannel.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { ChannelEnumChannelTypeEnum } from "./channelenumchanneltypeenum";


export class ChatV2ServiceChannel extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=attributes" })
  attributes?: string;

  @Metadata({ data: "json, name=created_by" })
  createdBy?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @Metadata({ data: "json, name=members_count" })
  membersCount?: number;

  @Metadata({ data: "json, name=messages_count" })
  messagesCount?: number;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=type" })
  type?: ChannelEnumChannelTypeEnum;

  @Metadata({ data: "json, name=unique_name" })
  uniqueName?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}